refactor(client): extract importable core
Move the client implementation into pkg/clientcore so mobile wrappers can call Run(ctx, Config) directly while the CLI remains a thin wrapper.

Adapted from ios-clientcore-refactor on top of the current wrap implementation.
