+++ pkg/clientcore/captcha_v2.go
@@ -1,4 +1,4 @@
-package main
+package clientcore
 
 import (
 	"bytes"
