+++ pkg/clientcore/manual_captcha.go
@@ -1,4 +1,4 @@
-package main
+package clientcore
 
 import (
 	"bytes"
