+++ pkg/clientcore/namegen.go
@@ -1,4 +1,4 @@
-package main
+package clientcore
 
 import (
 	"fmt"
