+++ pkg/clientcore/profiles.go
@@ -1,4 +1,4 @@
-package main
+package clientcore
 
 import (
 	"encoding/json"
