+++ pkg/clientcore/slider_captcha_test.go
@@ -1,4 +1,4 @@
-package main
+package clientcore
 
 import (
 	"bytes"
